refactor(scripts): share prompt helper
